task1/pyproject.toml
[tool.poetry]
name = "summer2023"
version = "0.1.2"
description = "Projection Pursuit, Machine Learning"
authors = ["David Zheng", "Tengyu Song", "Hongyan Zhou"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.2"
pandas = "^1.4.0"
hydra-core = "^1.3.1"
scikit-learn = "^1.0"
PyYAML = "^5.4"
jupyter = "^1.0"

[build-system]
requires = ["poetry-core>=0.12"]
build-backend = "poetry.core.masonry.api"